+++ apm-sniffer/apm-sdk-plugin/grpc-1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/grpc/v1/server/TracingServerCallListener.java
@@ -22,6 +22,7 @@ import io.grpc.ForwardingServerCallListener;
 import io.grpc.MethodDescriptor;
 import io.grpc.ServerCall;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
+import org.apache.skywalking.apm.agent.core.context.ContextSnapshot;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
 import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
@@ -34,11 +35,17 @@ import org.apache.skywalking.apm.plugin.grpc.v1.OperationNameFormatUtil;
 public class TracingServerCallListener<REQUEST> extends ForwardingServerCallListener.SimpleForwardingServerCallListener<REQUEST> {
     private final MethodDescriptor.MethodType methodType;
     private final String operationPrefix;
+    private final ContextSnapshot contextSnapshot;
+    private final AbstractSpan parentEntrySpan;
 
-    protected TracingServerCallListener(ServerCall.Listener<REQUEST> delegate, MethodDescriptor<REQUEST, ?> descriptor) {
+    protected TracingServerCallListener(ServerCall.Listener<REQUEST> delegate, MethodDescriptor<REQUEST, ?> descriptor,
+                                        final ContextSnapshot contextSnapshot,
+                                        final AbstractSpan parentEntrySpan) {
         super(delegate);
         this.methodType = descriptor.getType();
         this.operationPrefix = OperationNameFormatUtil.formatOperationName(descriptor) + SERVER;
+        this.contextSnapshot = contextSnapshot;
+        this.parentEntrySpan = parentEntrySpan;
     }
 
     @Override
@@ -48,7 +55,7 @@ public class TracingServerCallListener<REQUEST> extends ForwardingServerCallList
             final AbstractSpan span = ContextManager.createLocalSpan(operationPrefix + REQUEST_ON_MESSAGE_OPERATION_NAME);
             span.setComponent(ComponentsDefine.GRPC);
             span.setLayer(SpanLayer.RPC_FRAMEWORK);
-            ContextManager.continued(ServerInterceptor.CONTEXT_SNAPSHOT_KEY.get());
+            ContextManager.continued(contextSnapshot);
             try {
                 super.onMessage(message);
             } catch (Throwable t) {
@@ -67,7 +74,7 @@ public class TracingServerCallListener<REQUEST> extends ForwardingServerCallList
         final AbstractSpan span = ContextManager.createLocalSpan(operationPrefix + REQUEST_ON_CANCEL_OPERATION_NAME);
         span.setComponent(ComponentsDefine.GRPC);
         span.setLayer(SpanLayer.RPC_FRAMEWORK);
-        ContextManager.continued(ServerInterceptor.CONTEXT_SNAPSHOT_KEY.get());
+        ContextManager.continued(contextSnapshot);
         try {
             super.onCancel();
         } catch (Throwable t) {
@@ -75,7 +82,7 @@ public class TracingServerCallListener<REQUEST> extends ForwardingServerCallList
             throw t;
         } finally {
             ContextManager.stopSpan();
-            ServerInterceptor.ACTIVE_SPAN_KEY.get().asyncFinish();
+            parentEntrySpan.asyncFinish();
         }
     }
 
@@ -84,7 +91,7 @@ public class TracingServerCallListener<REQUEST> extends ForwardingServerCallList
         final AbstractSpan span = ContextManager.createLocalSpan(operationPrefix + REQUEST_ON_HALF_CLOSE_OPERATION_NAME);
         span.setComponent(ComponentsDefine.GRPC);
         span.setLayer(SpanLayer.RPC_FRAMEWORK);
-        ContextManager.continued(ServerInterceptor.CONTEXT_SNAPSHOT_KEY.get());
+        ContextManager.continued(contextSnapshot);
         try {
             super.onHalfClose();
         } catch (Throwable t) {